Update enhanced-text-widget.php
This fork adds functionality for line break (br) and span tags in the widget title. It's a self-confessed hack. Would be much better if a real PHP developer could run with this and made it more end-user friendly. (I.e. optional subheading line with optional span for differentiated styling.) As it is, it works for my own purposes, and maybe somebody else can use it too.